requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.1
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
evdev==1.6.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.113.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Session==0.8.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.8
importlib-metadata==4.6.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyboard==0.13.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
MouseInfo==0.1.3
msgspec==0.18.6
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.28.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.9.0
pydantic_core==2.23.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==2.4.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pystyle==2.9
pyTelegramBotAPI==4.22.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
PyYAML==5.4.1
pyzmq==26.2.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.4
systemd-python==234
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.30.6
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
yarg==0.1.9
yarl==1.17.1
zipp==1.0.0
